requirements.txt
astroid==2.4.2
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
paho-mqtt==1.5.0
psutil==5.7.0
pylint==2.5.3
PyYAML==5.3.1
six==1.15.0
toml==0.10.1
uptime==3.0.1
wrapt==1.12.1
